more typing work

+++ kitty/tabs.py
@@ -275,7 +275,7 @@ class Tab:  # {{{
         self._set_current_layout(layout_name)
         self.relayout()
 
-    def resize_window_by(self, window_id: int, increment: int, is_horizontal: bool) -> Optional[str]:
+    def resize_window_by(self, window_id: int, increment: float, is_horizontal: bool) -> Optional[str]:
         increment_as_percent = self.current_layout.bias_increment_for_cell(is_horizontal) * increment
         if self.current_layout.modify_size_of_window(self.windows, window_id, increment_as_percent, is_horizontal):
             self.relayout()
